Add a policy against dangling references in the registr
When discussing http://github.com/FamilySearch/GEDCOM.io/pull/285 the steering committee noted that it might make sense to require that if a YAML file with `translated from: de` is submitted to the registry then the registry must also have a file with the same URI with `lang: de`. This PR extends that idea to other kinds of dangling references (substructure, superstructure, enumset, and so on) and adds it to the README section on submitting new or changed YAML files.