Register census record extension used by MagiKey
This has been in use for many years in MagiKey applications
with GEDCOM 5.5.1 and 7.0.  I'm now trying to figure out how
to correctly register it.

These files were hand generated so I still need to do more
validation of their correctness, both in terms of YAML formatting,
and in terms of matching the implementation.

Signed-off-by: Dave Thaler <dthaler@armidalesoftware.com>