Check YAML paths and links between YAML files. (#183)
* Check YAML paths and links between YAML files.

This file could be added to github actions help with #180

At present, it fails because we have an error in some of the datatype files; https://github.com/FamilySearch/GEDCOM/issues/665 has been created to track that issue.

* add remaining keys

* typoe

* Update registry_tools/global-validator.py

Co-authored-by: Dave Thaler <dthaler1968@gmail.com>

* refactor to allow use as library

* Have global_validator.py use validator.py

Previously, global_validator.py could crash if validator.py was not run first; this change fixes that. It also removes the necessity for the workflow to use bash to pick the files to validate or to run validator.py directly, as global_validator.py now does both operations.

* refactor validator into two files

---------

Co-authored-by: Dave Thaler <dthaler1968@gmail.com>