Register manifest for MagiKey census record extension
PR #181 registered the YAML files but not a manifest.
This PR adds a manifest for that set of related files,
as requested in issue #179.

An interesting case of note is that the same tag (`_CENR`)
has two rows, one for the record YAML and one for the
substructure YAML.  This happens in the standard manifests
too, but there's no `tag` column in the standard manifests.

Signed-off-by: Dave Thaler <dthaler@armidalesoftware.com>